+++ common/src/main/java/dev/terminalmc/nocapes/mixin/MixinCapeLayer.java
@@ -26,13 +26,28 @@ import net.minecraft.client.renderer.entity.layers.CapeLayer;
 import net.minecraft.world.entity.player.PlayerModelPart;
 import org.spongepowered.asm.mixin.Mixin;
 
+import static dev.terminalmc.nocapes.config.Config.options;
+
 @Mixin(CapeLayer.class)
 public class MixinCapeLayer {
-    @WrapMethod(method = "render(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;ILnet/minecraft/client/player/AbstractClientPlayer;FFFFFF)V")
-    private void wrapRender(PoseStack poseStack, MultiBufferSource buffer, int packedLight, AbstractClientPlayer player, float limbSwing, float limbSwingAmount, float partialTicks, float ageInTicks, float netHeadYaw, float headPitch, Operation<Void> original) {
+    @WrapMethod(
+            method = "render(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;ILnet/minecraft/client/player/AbstractClientPlayer;FFFFFF)V"
+    )
+    private void wrapRender(PoseStack poseStack, MultiBufferSource buffer, int packedLight, 
+                            AbstractClientPlayer player, float limbSwing, float limbSwingAmount, 
+                            float partialTicks, float ageInTicks, float netHeadYaw, float headPitch, 
+                            Operation<Void> original) {
         // Skip the check if the cape wouldn't be rendered anyway
-        if (!player.isInvisible() && player.isModelPartShown(PlayerModelPart.CAPE) && !NoCapes.blockCape(player.getGameProfile())) {
-            original.call(poseStack, buffer, packedLight, player, limbSwing, limbSwingAmount, partialTicks, ageInTicks, netHeadYaw, headPitch);
+        if (
+                !player.isInvisible() 
+                && player.isModelPartShown(PlayerModelPart.CAPE) 
+                && (
+                        !options().hideCape 
+                        || !NoCapes.blockCape(player.getGameProfile())
+                )
+        ) {
+            original.call(poseStack, buffer, packedLight, player, limbSwing, limbSwingAmount, 
+                    partialTicks, ageInTicks, netHeadYaw, headPitch);
         }
     }
 }