+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.1
+
+- Added Mojang Office cape
+- Added options to individually switch cape and elytra off
+
 ## 1.0.0
 
 - Initial full release